+++ insta/src/runtime.rs
@@ -32,6 +32,7 @@ thread_local! {
 
 // This macro is basically eprintln but without being captured and
 // hidden by the test runner.
+#[macro_export]
 macro_rules! elog {
     () => (write!(std::io::stderr()).ok());
     ($($arg:tt)*) => ({
@@ -49,38 +50,38 @@ pub struct AutoName;
 
 impl From<AutoName> for ReferenceValue<'static> {
     fn from(_value: AutoName) -> ReferenceValue<'static> {
-        ReferenceValue::Named(None)
+        ReferenceValue::File(None)
     }
 }
 
 impl From<Option<String>> for ReferenceValue<'static> {
     fn from(value: Option<String>) -> ReferenceValue<'static> {
-        ReferenceValue::Named(value.map(Cow::Owned))
+        ReferenceValue::File(value.map(Cow::Owned))
     }
 }
 
 impl From<String> for ReferenceValue<'static> {
     fn from(value: String) -> ReferenceValue<'static> {
-        ReferenceValue::Named(Some(Cow::Owned(value)))
+        ReferenceValue::File(Some(Cow::Owned(value)))
     }
 }
 
 impl<'a> From<Option<&'a str>> for ReferenceValue<'a> {
     fn from(value: Option<&'a str>) -> ReferenceValue<'a> {
-        ReferenceValue::Named(value.map(Cow::Borrowed))
+        ReferenceValue::File(value.map(Cow::Borrowed))
     }
 }
 
 impl<'a> From<&'a str> for ReferenceValue<'a> {
     fn from(value: &'a str) -> ReferenceValue<'a> {
-        ReferenceValue::Named(Some(Cow::Borrowed(value)))
+        ReferenceValue::File(Some(Cow::Borrowed(value)))
     }
 }
 
 /// A reference to a snapshot
 pub enum ReferenceValue<'a> {
-    /// A named snapshot, where the inner value is the snapshot name.
-    Named(Option<Cow<'a, str>>),
+    /// A file snapshot, where the inner value is the snapshot name.
+    File(Option<Cow<'a, str>>),
     /// An inline snapshot, where the inner value is the snapshot contents.
     Inline(&'a str),
 }
@@ -197,6 +198,8 @@ fn get_snapshot_filename(
     })
 }
 
+/// A single snapshot including surrounding context which asserts and save the
+/// snapshot.
 #[derive(Debug)]
 struct SnapshotAssertionContext<'a> {
     tool_config: Arc<ToolConfig>,
@@ -231,7 +234,7 @@ impl<'a> SnapshotAssertionContext<'a> {
         let is_doctest = is_doctest(function_name);
 
         match refval {
-            ReferenceValue::Named(name) => {
+            ReferenceValue::File(name) => {
                 let name = match name {
                     Some(name) => add_suffix_to_snapshot_name(name),
                     None => {
@@ -368,6 +371,17 @@ impl<'a> SnapshotAssertionContext<'a> {
         let should_print = self.tool_config.output_behavior() != OutputBehavior::Nothing;
         let snapshot_update = snapshot_update_behavior(&self.tool_config, unseen);
 
+        // If snapshot_update is `InPlace` and we have an inline snapshot, then
+        // use `NewFile`, since we can't use `InPlace` for inline. `cargo-insta`
+        // then accepts all snapshots at the end of the test.
+
+        let snapshot_update =
+            if snapshot_update == SnapshotUpdateBehavior::InPlace && self.snapshot_file.is_none() {
+                SnapshotUpdateBehavior::NewFile
+            } else {
+                snapshot_update
+            };
+
         match snapshot_update {
             SnapshotUpdateBehavior::InPlace => {
                 if let Some(ref snapshot_file) = self.snapshot_file {
@@ -383,20 +397,14 @@ impl<'a> SnapshotAssertionContext<'a> {
                             style(snapshot_file.display()).cyan().underlined(),
                         );
                     }
-                } else if should_print {
-                    elog!(
-                        "{}",
-                        style(
-                            "error: cannot update inline snapshots in-place \
-                        (https://github.com/mitsuhiko/insta/issues/272)"
-                        )
-                        .red()
-                        .bold(),
-                    );
+                } else {
+                    // Checked self.snapshot_file.is_none() above
+                    unreachable!()
                 }
             }
             SnapshotUpdateBehavior::NewFile => {
                 if let Some(ref snapshot_file) = self.snapshot_file {
+                    // File snapshot
                     if let Some(new_path) = new_snapshot.save_new(snapshot_file)? {
                         if should_print {
                             elog!(
@@ -415,19 +423,7 @@ impl<'a> SnapshotAssertionContext<'a> {
                                 .bold(),
                         );
                     }
-
-                // special case for pending inline snapshots.  Here we really only want
-                // to write the contents if the snapshot contents changed as the metadata
-                // is not retained for inline snapshots.  This used to have different
-                // behavior in the past where we did indeed want to rewrite the snapshots
-                // entirely since we used to change the canonical snapshot format, but now
-                // this is significantly less likely to happen and seeing hundreds of unchanged
-                // inline snapshots in the review screen is not a lot of fun.
-                } else if self
-                    .old_snapshot
-                    .as_ref()
-                    .map_or(true, |x| x.contents() != new_snapshot.contents())
-                {
+                } else {
                     PendingInlineSnapshot::new(
                         Some(new_snapshot),
                         self.old_snapshot.clone(),
@@ -678,7 +674,10 @@ pub fn assert_snapshot(
     if pass {
         ctx.cleanup_passing()?;
 
-        if tool_config.force_update_snapshots() {
+        if matches!(
+            tool_config.snapshot_update(),
+            crate::env::SnapshotUpdate::Force
+        ) {
             ctx.update_snapshot(new_snapshot)?;
         }
     // otherwise print information and update snapshots.